feat: simplify configs
BREAKING CHANGE: ES6, import & main configs don't have sub-configs anymore
