--- es6/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  extends: [
-    './non-rules-config.js',
-    './stylistic.js',
-    './best-practices.js',
-    './possible-errors.js',
-  ],
-  rules: {},
-}
