--- es6/possible-errors.js
@@ -1,31 +1,0 @@
-module.exports = {
-  extends: './non-rules-config.js',
-  rules: {
-    'constructor-super': 'error',
-    'no-const-assign': 'error',
-    'no-dupe-class-members': 'error',
-    'no-new-symbol': 'error',
-    'no-this-before-super': 'error',
-    'no-unsafe-optional-chaining': 'error',
-
-    'no-unused-expressions': 'off',
-    '@babel/no-unused-expressions': 'error',
-  },
-  overrides: [
-    {
-      files: ['**/*.ts?(x)'],
-      rules: {
-        'constructor-super': 'off', // ts(2335) & ts(2377)
-        'no-const-assign': 'off', // ts(2588)
-        'no-new-symbol': 'off', // ts(2588)
-        'no-this-before-super': 'off', // ts(2376)
-
-        'no-dupe-class-members': 'off',
-        '@typescript-eslint/no-dupe-class-members': 'off', // ts(2393) & ts(2300)
-
-        '@babel/no-unused-expressions': 'off',
-        '@typescript-eslint/no-unused-expressions': 'error',
-      },
-    },
-  ],
-}
