--- import/deprecated-rules.js
@@ -1,6 +1,0 @@
-const {mapDeprecatedRules} = require('../utils')
-
-module.exports = {
-  extends: './non-rules-config.js',
-  rules: mapDeprecatedRules(['import/imports-first']),
-}
