--- import/index.js
@@ -1,11 +1,0 @@
-module.exports = {
-  extends: [
-    './non-rules-config.js',
-    './deprecated-rules.js',
-    './stylistic.js',
-    './best-practices.js',
-    './possible-errors.js',
-    './disabled-rules.js',
-  ],
-  rules: {},
-}
