--- import/non-rules-config.js
@@ -1,26 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-  },
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module',
-  },
-  plugins: ['import'],
-  settings: {
-    'import/ignore': ['node_modules', '.json$', '.(scss|less|css|styl)$'],
-  },
-  overrides: [
-    {
-      files: ['**/*.ts?(x)'],
-      extends: 'plugin:import/typescript',
-      parser: '@typescript-eslint/parser',
-      parserOptions: {
-        ecmaVersion: 2018,
-        sourceType: 'module',
-      },
-      plugins: ['@typescript-eslint'],
-      rules: {},
-    },
-  ],
-}
