--- non-rules-config.js
@@ -1,29 +1,0 @@
-const fs = require('fs')
-const path = require('path')
-
-const tsConfig = fs.existsSync('tsconfig.json')
-  ? path.resolve('tsconfig.json')
-  : fs.existsSync('types/tsconfig.json')
-  ? path.resolve('types/tsconfig.json')
-  : undefined
-
-module.exports = {
-  env: {
-    browser: true,
-    node: true,
-  },
-  rules: {},
-  overrides: [
-    {
-      files: ['**/*.ts?(x)'],
-      parser: '@typescript-eslint/parser',
-      parserOptions: {
-        ecmaVersion: 2018,
-        sourceType: 'module',
-        project: tsConfig,
-      },
-      plugins: ['@typescript-eslint'],
-      rules: {},
-    },
-  ],
-}
